// video-app/apps/room-service/src/common/enums.ts
export enum RoomType {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
}

export enum RoomStatus {
  ACTIVE = 'ACTIVE',
  CLOSED = 'CLOSED',
}

export enum MemberRole {
  OWNER = 'OWNER',
  MODERATOR = 'MODERATOR',
  PARTICIPANT = 'PARTICIPANT',
}

export enum MemberStatus {
  JOINED = 'JOINED',
  LEFT = 'LEFT',
  KICKED = 'KICKED',
}